requirements.txt
openai
python-dotenv
pymongo==4.6.1
redis==4.6.0
psutil
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6
motor==3.3.2
sqlite3
asyncio
uuid
datetime
typing
pathlib
logging
json
re
PyJWT
